raster same and staircase call macros
