+++ datatypes/src/util/mod.rs
@@ -2,5 +2,6 @@ pub mod arrow;
 pub mod helpers;
 pub mod ranges;
 mod result;
+pub mod test;
 
 pub use result::Result;
